requirements.txt
django
psycopg2-binary
djangorestframework
django-environ
gunicorn
celery
redis
telethon
daphne
whitenoise[brotli]
